requirements.txt
pyTelegramBotAPI==4.19.2
python-dotenv==1.0.0
